Task 15.2 aggregator

+++ Task_15.2/src/by/epam/aggregator/ResultSetter.java
@@ -4,5 +4,6 @@ package by.epam.aggregator;
  * Created by Olga Shahray on 06.07.2016.
  */
 public interface ResultSetter {
-     void setResult(Float result);
+
+     void setResult(Double result);
 }
